Move files to reduce nesting
