+++ app/utils/mdx.ts
@@ -2,14 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import matter from 'gray-matter';
 import markdownit from 'markdown-it';
-import { transformToDatasetsList, processTaxonomies } from './data';
-import type {
-  DatasetMetadata,
-  ContentMetadata,
-  DatasetWithContent,
-  StoryWithContent,
-  StoryMetadata,
-} from 'app/types/content';
+
 import { compileMDX } from 'next-mdx-remote/rsc';
 import React from 'react';
 import { Paragraph } from 'app/components/common/Paragraph';
@@ -30,19 +23,6 @@ type ThemeFrontmatter = {
   image: string;
   pubDate?: string;
 };
-
-const STORY_CONTENT_PATH = path.join(
-  process.cwd(),
-  'app',
-  'content',
-  'stories',
-);
-const DATASET_CONTENT_PATH = path.join(
-  process.cwd(),
-  'app',
-  'content',
-  'datasets',
-);
 
 const THEME_CONTENT_PATH = path.join(process.cwd(), 'app', 'content', 'themes');
 
@@ -102,60 +82,6 @@ export function readMDXFile(filePath) {
   const rawContent = fs.readFileSync(filePath, 'utf-8');
   const parsedData = matter(rawContent);
   return parsedData;
-}
-
-function getMDXData(dir): ContentMetadata[] {
-  const mdxFiles = getMDXFiles(dir);
-  return mdxFiles.map((file) => {
-    const { content, data } = readMDXFile(path.join(dir, file));
-    const parsedData = parseAttributes(data);
-    const processedData = processTaxonomies(parsedData);
-    const slug = path.basename(file, path.extname(file));
-
-    return {
-      metadata: processedData,
-      slug,
-      content,
-    };
-  });
-}
-
-function getMDXMetaData(dir: string): ContentMetadata[] {
-  const mdxFiles = getMDXFiles(dir);
-  return mdxFiles.map((file) => {
-    const { data } = readMDXFile(path.join(dir, file));
-    const parsedData = parseAttributes(data);
-    const processedData = processTaxonomies(parsedData);
-    const slug = path.basename(file, path.extname(file));
-    return {
-      metadata: processedData,
-      slug,
-    };
-  });
-}
-
-export function getStoriesMetadata(): StoryMetadata[] {
-  return getMDXMetaData(STORY_CONTENT_PATH) as StoryMetadata[];
-}
-
-export function getStories() {
-  return getMDXData(STORY_CONTENT_PATH) as StoryWithContent[];
-}
-
-export function getDatasetsMetadata(): DatasetMetadata[] {
-  return getMDXMetaData(DATASET_CONTENT_PATH) as DatasetMetadata[];
-}
-
-export function getDatasets(): DatasetWithContent[] {
-  return getMDXData(DATASET_CONTENT_PATH) as DatasetWithContent[];
-}
-
-export function getTransformedDatasetMetadata() {
-  return transformToDatasetsList(getDatasetsMetadata());
-}
-
-export function getTransformedDatasets() {
-  return transformToDatasetsList(getDatasets());
 }
 
 /**
